add comments and simplify

+++ bodo/tests/test_nested_array.cpp
@@ -117,8 +117,9 @@ std::shared_ptr<array_info> get_sample_struct(size_t nfields = 1) {
     // Create the array item array with one child array
     auto struct_arr = alloc_struct(10, children);
     // Name the fields alphabetically
-    for (uint64_t i = 0; i < static_cast<uint64_t>(nfields); ++i) {
-        struct_arr->field_names.emplace_back(i / 26 + 1, 'a' + i % 26);
+    for (size_t i = 0; i < (nfields); ++i) {
+        struct_arr->field_names.emplace_back(static_cast<uint64_t>(i / 26 + 1),
+                                             'a' + i % 26);
     }
     return struct_arr;
 }
